+++ test/test-providers/main.ts
@@ -5,13 +5,10 @@ import * as Aws from "./.gen/providers/aws";
 import * as Azure from "./.gen/providers/azurerm";
 import * as Google from "./.gen/providers/google";
 import * as Kubernetes from "./.gen/providers/kubernetes";
-// The following providers still have bugs
-// - https://github.com/hashicorp/terraform-cdk/issues/125
-// - https://github.com/hashicorp/terraform-cdk/issues/124
-//
-// import * as Nomad from "./.gen/providers/nomad";
-// import * as Vault from "./.gen/providers/vault";
-// import * as Consul from "./.gen/providers/consul";
+import * as Openstack from "./.gen/providers/openstack";
+import * as Nomad from "./.gen/providers/nomad";
+import * as Vault from "./.gen/providers/vault";
+import * as Consul from "./.gen/providers/consul";
 
 export class HelloTerra extends TerraformStack {
   constructor(scope: Construct, id: string) {
@@ -27,7 +24,7 @@ export class HelloTerra extends TerraformStack {
       },
     ]);
 
-    [Aws, Azure, Google, Kubernetes];
+    [Aws, Azure, Google, Kubernetes, Nomad, Vault, Openstack, Consul];
   }
 }